Move 'GetUserNameFromPasswd' related code from Environment.Unix.cs to Interop.GetPwUid.cs
